Begin post login front end development, reorganize file structure of front end components
