+++ stonks_frontend/src/components/pages/Dashboard.js
@@ -1,0 +1,18 @@
+import { StatusBar } from 'expo-status-bar';
+import React, { Component } from 'react';
+import { StyleSheet, View, Image, Text, Button } from 'react-native';
+
+/** What is to be the MAIN PAGE of the User experience on mobile once logged in.
+ *    Will ultimately have a NAVIGATION DRAWER (according to Nanc) to help navigate around
+ *      our pages and services. Contents of this page are up for discussion upon the next 
+ *      all hands meeting. 
+ */
+
+class Dashboard extends Component {
+
+    render() {
+        return(<View></View>);
+    }
+}
+
+export default Dashboard;
